+++ NEWS.md
@@ -1,7 +1,8 @@
 # climate 1.2.0
 
-* Support for telemetry dataset from IMGW-PIB repository - currently only for air temperature
-* update for location of IMGW-PIB stations
+* Support for telemetry dataset from IMGW-PIB repository
+    * `data.table` package used for efficient processing of large amount of data in this subrepository
+* Update for location of IMGW-PIB telemetry stations
 
 # climate 1.1.1