+++ NEWS.md
@@ -4,6 +4,7 @@
     * `data.table` package used for efficient processing of large amount of data in this subrepository
 * Update for location of IMGW-PIB telemetry stations
 
+
 # climate 1.1.1
 
 * Fix problems with downloading `precip` dataset from IMGW-PIB repository after recent changes in metadata